% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distributionalDistContinuous}
\alias{distributionalDistContinuous}
\title{Distributional distance computation for continuous varialbles '}
\usage{
distributionalDistContinuous(X1, X2, nM)
}
\arguments{
\item{X1}{The first serie of integers}

\item{X2}{The second serie of integers}

\item{nM}{a vector of length 2 with the initial values for the first discretization. If missing will be filled with min and max values}
}
\value{
The distributional disancte between X1 and X2 using a $2^-i$ ponderation
}
\description{
Distributional distance computation for continuous varialbles '
}
\examples{
X1 <- c(-1,2.6,-4,5.5,7,1.8 );
X2 <- c(-0.5,2.6,-4,5.5,9,1.81);
distributionalDistContinuous(X1, X2)
}
