% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distributionalDistDiscrete}
\alias{distributionalDistDiscrete}
\title{Distributional distance computation for discrete varialbles '}
\usage{
distributionalDistDiscrete(X1, X2)
}
\arguments{
\item{X1}{The first serie of integers}

\item{X2}{The second serie of integers}
}
\value{
The distributional disancte between X1 and X2 using a $2^-i$ ponderation
}
\description{
Distributional distance computation for discrete varialbles '
}
\examples{
X1 <- as.integer(c(12, 13, 14, 12, 13, 16, 12, 13, 15, 12, 13, 16, 11, 13, 15, 12, 13, 16))
X2 <- as.integer(c(12, 13, 14, 19, 13, 16, 12, 13, 15, 12, 13, 16, 11, 13, 15, 12, 13, 16))
distributionalDistDiscrete(X1, X2)
}
